requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.100.0
sqlalchemy>=2.0.36
# Додаємо greenlet явно, оскільки SQLAlchemy його потребує для async
greenlet>=3.1.1
requests==2.31.0
aiohttp==3.9.1
python-dateutil==2.8.2
Pillow>=10.2.0
gtfs-realtime-bindings==1.0.0
rapidfuzz==3.5.2
cachetools==5.3.2
aiosqlite==0.19.0
redis>=4.5.0              # Для кешування
asyncpg>=0.28.0          # Для PostgreSQL
tenacity>=8.2.0          # Для retry логіки
structlog>=23.1.0        # Для структурованого логування
psutil>=5.9.0
psycopg2-binary
sqlalchemy[asyncio]